requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7

# PDF processing
PyPDF2==3.0.1
PyMuPDF==1.23.8
Pillow==10.1.0
reportlab==4.0.7

# Image processing
opencv-python-headless==4.8.1.78
numpy==1.24.3

# Utilities
aiofiles==23.2.1
python-multipart==0.0.6
requests==2.31.0
psutil==5.9.6

# Data handling
orjson==3.9.10
pydantic==2.5.0

# Async support
aiohttp==3.9.1
aioredis==2.0.1

# Logging and monitoring
structlog==23.2.0

# Environment
python-dotenv==1.0.0

# Date/time handling
python-dateutil==2.8.2

# File compression
zipfile36==0.1.3

# Text processing
chardet==5.2.0

# Security
cryptography==41.0.7